Refresh children segment views when parent changes
When there are two levels of segments (0 and 1), sometimes, the parent
view is reloaded but the second level view is not rendered after
reloading the parent; this results in the second level view just being
missing.

This commit should fix this issue by reloading children segments when
the parent segment itself is reloaded.

So far, I have not been able to find any side-effect in this change,
and it fixes my issue, but YMMV. Fixes issue #70.